Refactor card UI into modular views
