+++ src/ui/cards/shared.js
@@ -1,0 +1,413 @@
+import elements from '../elements.js';
+import { formatHours, formatMoney } from '../../core/helpers.js';
+import {
+  describeRequirement,
+  getDefinitionRequirements
+} from '../../game/requirements.js';
+import {
+  assignInstanceToNiche,
+  getAssignableNicheSummaries
+} from '../../game/assets/niches.js';
+import {
+  canPerformQualityAction,
+  getQualityActionUsage,
+  getQualityActions,
+  getQualityLevelSummary,
+  getQualityTracks,
+  performQualityAction
+} from '../../game/assets/quality.js';
+
+export function showSlideOver({ eyebrow, title, body }) {
+  const { slideOver, slideOverContent, slideOverEyebrow, slideOverTitle } = elements;
+  if (!slideOver || !slideOverContent) return;
+  slideOverEyebrow.textContent = eyebrow || '';
+  slideOverTitle.textContent = title || '';
+  slideOverContent.innerHTML = '';
+  if (Array.isArray(body)) {
+    body.forEach(node => node && slideOverContent.appendChild(node));
+  } else if (body instanceof Node) {
+    slideOverContent.appendChild(body);
+  } else if (typeof body === 'string') {
+    const p = document.createElement('p');
+    p.textContent = body;
+    slideOverContent.appendChild(p);
+  }
+  slideOver.hidden = false;
+  slideOver.focus();
+}
+
+export function createBadge(label) {
+  const span = document.createElement('span');
+  span.className = 'badge';
+  span.textContent = label;
+  return span;
+}
+
+export function createDefinitionSummary(title, rows = []) {
+  const section = document.createElement('section');
+  const heading = document.createElement('h3');
+  heading.textContent = title;
+  section.appendChild(heading);
+  const list = document.createElement('ul');
+  list.className = 'definition-list';
+  rows.forEach(row => {
+    const item = document.createElement('li');
+    if (row.label) {
+      const label = document.createElement('span');
+      label.textContent = row.label;
+      label.className = 'definition-list__label';
+      item.appendChild(label);
+    }
+    if (row.value) {
+      const value = document.createElement('span');
+      if (row.value instanceof Node) {
+        value.appendChild(row.value);
+      } else {
+        value.textContent = row.value;
+      }
+      value.className = 'definition-list__value';
+      item.appendChild(value);
+    }
+    list.appendChild(item);
+  });
+  section.appendChild(list);
+  return section;
+}
+
+export function createAssetDetailHighlights(definition) {
+  const entries = Array.isArray(definition.detailEntries)
+    ? definition.detailEntries
+    : Array.isArray(definition.details)
+      ? definition.details.map((render, index) => ({ key: `detail-${index}`, render }))
+      : [];
+
+  const renderedDetails = entries
+    .map((entry, index) => {
+      const render = typeof entry.render === 'function' ? entry.render : entry;
+      if (typeof render !== 'function') return null;
+      try {
+        const value = render();
+        if (!value) return null;
+        if (typeof value === 'string') {
+          const trimmed = value.trim();
+          if (!trimmed) return null;
+          return { key: entry.key || `detail-${index}`, value: trimmed };
+        }
+        if (value instanceof Node) {
+          return { key: entry.key || `detail-${index}`, value };
+        }
+        return null;
+      } catch (error) {
+        console.error('Failed to render asset detail', error);
+        return null;
+      }
+    })
+    .filter(Boolean);
+
+  const detailByKey = new Map();
+  renderedDetails.forEach(detail => {
+    if (!detailByKey.has(detail.key)) {
+      detailByKey.set(detail.key, []);
+    }
+    detailByKey.get(detail.key).push(detail.value);
+  });
+
+  const requirements = getDefinitionRequirements(definition);
+  const section = document.createElement('section');
+  section.className = 'asset-detail__section asset-detail__section--blueprint';
+  const heading = document.createElement('h3');
+  heading.textContent = 'Launch blueprint';
+  section.appendChild(heading);
+
+  const grid = document.createElement('div');
+  grid.className = 'asset-detail__summary-grid';
+  section.appendChild(grid);
+
+  const requirementsCard = document.createElement('article');
+  requirementsCard.className = 'asset-detail__summary-card asset-detail__summary-card--requirements';
+  const requirementsTitle = document.createElement('h4');
+  requirementsTitle.textContent = 'Requirements';
+  requirementsCard.appendChild(requirementsTitle);
+  const requirementsList = document.createElement('ul');
+  requirementsList.className = 'asset-detail__summary-list';
+  if (requirements?.hasAny) {
+    requirements.all.forEach(req => {
+      const descriptor = describeRequirement(req);
+      const item = document.createElement('li');
+      item.className = 'asset-detail__summary-item';
+      item.innerHTML = descriptor?.detail || '❔ <strong>Requirement</strong>';
+      requirementsList.appendChild(item);
+    });
+  } else {
+    const fallbackItem = document.createElement('li');
+    fallbackItem.className = 'asset-detail__summary-item';
+    const fallbackDetail = detailByKey.get('requirements')?.[0] || '🔓 Requirements: <strong>None</strong>';
+    if (typeof fallbackDetail === 'string') {
+      fallbackItem.innerHTML = fallbackDetail;
+    } else if (fallbackDetail instanceof Node) {
+      fallbackItem.appendChild(fallbackDetail);
+    }
+    requirementsList.appendChild(fallbackItem);
+  }
+  requirementsCard.appendChild(requirementsList);
+  grid.appendChild(requirementsCard);
+
+  const roadmapCard = document.createElement('article');
+  roadmapCard.className = 'asset-detail__summary-card asset-detail__summary-card--roadmap';
+  const roadmapTitle = document.createElement('h4');
+  roadmapTitle.textContent = 'Roadmap & stats';
+  roadmapCard.appendChild(roadmapTitle);
+
+  const roadmapList = document.createElement('ul');
+  roadmapList.className = 'asset-detail__summary-list';
+  const roadmapKeys = ['owned', 'setup', 'setupCost', 'maintenance', 'income', 'latestYield'];
+  roadmapKeys.forEach(key => {
+    const values = detailByKey.get(key) || [];
+    values.forEach(value => {
+      const item = document.createElement('li');
+      item.className = 'asset-detail__summary-item';
+      if (typeof value === 'string') {
+        item.innerHTML = value;
+      } else if (value instanceof Node) {
+        item.appendChild(value);
+      }
+      roadmapList.appendChild(item);
+    });
+  });
+
+  const consumedKeys = new Set(['requirements', 'qualitySummary', 'qualityProgress', ...roadmapKeys]);
+  const extraDetails = renderedDetails.filter(detail => !consumedKeys.has(detail.key));
+  extraDetails.forEach(detail => {
+    const item = document.createElement('li');
+    item.className = 'asset-detail__summary-item';
+    if (typeof detail.value === 'string') {
+      item.innerHTML = detail.value;
+    } else if (detail.value instanceof Node) {
+      item.appendChild(detail.value);
+    }
+    roadmapList.appendChild(item);
+  });
+
+  if (!roadmapList.children.length) {
+    const empty = document.createElement('li');
+    empty.className = 'asset-detail__summary-item';
+    empty.textContent = 'No roadmap details available yet.';
+    roadmapList.appendChild(empty);
+  }
+
+  roadmapCard.appendChild(roadmapList);
+  grid.appendChild(roadmapCard);
+
+  const qualityCard = document.createElement('article');
+  qualityCard.className = 'asset-detail__summary-card asset-detail__summary-card--quality';
+  const qualityTitle = document.createElement('h4');
+  qualityTitle.textContent = 'Quality journey';
+  qualityCard.appendChild(qualityTitle);
+
+  const qualitySummary = detailByKey.get('qualitySummary')?.[0];
+  const summaryCopy = document.createElement('p');
+  summaryCopy.className = 'asset-detail__summary-copy';
+  if (typeof qualitySummary === 'string') {
+    summaryCopy.innerHTML = qualitySummary;
+  } else if (qualitySummary instanceof Node) {
+    summaryCopy.appendChild(qualitySummary);
+  } else {
+    summaryCopy.textContent = '✨ Quality boosts unlock as you invest in specialty tracks.';
+  }
+  qualityCard.appendChild(summaryCopy);
+
+  const qualityList = document.createElement('ul');
+  qualityList.className = 'asset-detail__summary-list asset-detail__summary-list--quality';
+  const tracks = getQualityTracks(definition);
+  const levels = getQualityLevelSummary(definition);
+  levels.forEach(level => {
+    const item = document.createElement('li');
+    item.className = 'asset-detail__summary-item';
+    const title = document.createElement('div');
+    title.className = 'asset-detail__summary-line';
+    title.innerHTML = `<strong>Quality ${level.level}:</strong> ${level.name}`;
+    item.appendChild(title);
+    const requirementEntries = Object.entries(level.requirements || {});
+    if (requirementEntries.length) {
+      const detail = document.createElement('div');
+      detail.className = 'asset-detail__summary-subtext';
+      const parts = requirementEntries.map(([key, value]) => {
+        const label = tracks[key]?.shortLabel || tracks[key]?.label || key;
+        return `${value} ${label}`;
+      });
+      detail.textContent = parts.join(' • ');
+      item.appendChild(detail);
+    } else {
+      const detail = document.createElement('div');
+      detail.className = 'asset-detail__summary-subtext';
+      detail.textContent = 'Entry tier — no prep required.';
+      item.appendChild(detail);
+    }
+    qualityList.appendChild(item);
+  });
+  qualityCard.appendChild(qualityList);
+  grid.appendChild(qualityCard);
+
+  return section;
+}
+
+export function createInstanceQuickActions(definition, instance, state) {
+  const container = document.createElement('div');
+  container.className = 'asset-detail__quick-actions';
+
+  if (instance.status !== 'active') {
+    const note = document.createElement('span');
+    note.className = 'asset-detail__action-note';
+    note.textContent = 'Upgrades unlock after launch.';
+    container.appendChild(note);
+    return container;
+  }
+
+  const actions = getQualityActions(definition);
+  if (!actions.length) {
+    const note = document.createElement('span');
+    note.className = 'asset-detail__action-note';
+    note.textContent = 'No quality actions configured yet.';
+    container.appendChild(note);
+    return container;
+  }
+
+  const prioritized = [...actions].sort((a, b) => {
+    const aAvailable = canPerformQualityAction(definition, instance, a, state) ? 1 : 0;
+    const bAvailable = canPerformQualityAction(definition, instance, b, state) ? 1 : 0;
+    return bAvailable - aAvailable;
+  });
+
+  const limit = Math.min(prioritized.length, 3);
+  for (let index = 0; index < limit; index += 1) {
+    const action = prioritized[index];
+    if (!action) continue;
+    const button = document.createElement('button');
+    button.type = 'button';
+    button.className = 'asset-detail__action-button';
+    button.textContent = action.label || 'Upgrade';
+    const disabled = !canPerformQualityAction(definition, instance, action, state);
+    button.disabled = disabled;
+    const details = [];
+    if (action.time) {
+      details.push(`⏳ ${formatHours(action.time)}`);
+    }
+    if (action.cost) {
+      details.push(`💵 $${formatMoney(action.cost)}`);
+    }
+    const usage = getQualityActionUsage(definition, instance, action);
+    if (usage.dailyLimit > 0) {
+      details.push(`🔁 ${usage.remainingUses}/${usage.dailyLimit} today`);
+    }
+    let tooltip = details.join(' · ');
+    if (usage.exhausted) {
+      tooltip = `${tooltip ? `${tooltip} · ` : ''}All uses spent today. Come back tomorrow for a fresh charge.`;
+    }
+    if (tooltip) {
+      button.title = tooltip;
+    }
+    button.addEventListener('click', event => {
+      event.preventDefault();
+      if (button.disabled) return;
+      performQualityAction(definition.id, instance.id, action.id);
+    });
+    container.appendChild(button);
+  }
+
+  if (prioritized.length > limit) {
+    const more = document.createElement('span');
+    more.className = 'asset-detail__action-note';
+    more.textContent = `+${prioritized.length - limit} more upgrades available`;
+    container.appendChild(more);
+  }
+
+  return container;
+}
+
+export function createInstanceNicheSelector(definition, instance) {
+  if (!definition || !instance) return null;
+  const summaries = getAssignableNicheSummaries(definition);
+  if (!summaries.length) return null;
+
+  const container = document.createElement('div');
+  container.className = 'asset-detail__niche-selector';
+
+  const label = document.createElement('label');
+  label.className = 'asset-detail__niche-label';
+  const selectId = `asset-niche-${instance.id}`;
+  label.setAttribute('for', selectId);
+  label.textContent = 'Target niche';
+  container.appendChild(label);
+
+  const select = document.createElement('select');
+  select.className = 'asset-detail__niche-dropdown';
+  select.id = selectId;
+  select.name = selectId;
+
+  const placeholder = document.createElement('option');
+  placeholder.value = '';
+  placeholder.textContent = 'Unassigned';
+  select.appendChild(placeholder);
+
+  summaries.forEach(entry => {
+    if (!entry?.definition) return;
+    const option = document.createElement('option');
+    option.value = entry.definition.id;
+    const labelParts = [entry.definition.name];
+    if (entry.popularity?.label) {
+      labelParts.push(entry.popularity.label);
+    }
+    option.textContent = labelParts.join(' • ');
+    select.appendChild(option);
+  });
+
+  const currentValue = typeof instance.nicheId === 'string' ? instance.nicheId : '';
+  select.value = currentValue && Array.from(select.options).some(opt => opt.value === currentValue)
+    ? currentValue
+    : '';
+
+  const hint = document.createElement('p');
+  hint.className = 'asset-detail__niche-note';
+
+  function updateHint(selectedId) {
+    const match = summaries.find(entry => entry?.definition?.id === selectedId) || null;
+    if (!match || !match.popularity) {
+      hint.textContent = 'Choose a niche to tap into daily popularity rerolls.';
+      return;
+    }
+    const multiplier = Number(match.popularity.multiplier);
+    let percentLabel = '±0%';
+    if (Number.isFinite(multiplier)) {
+      const percent = Math.round((multiplier - 1) * 100);
+      const sign = percent > 0 ? '+' : '';
+      percentLabel = `${sign}${percent}%`;
+    }
+    const summary = match.popularity.summary || 'Demand shifts update daily.';
+    hint.textContent = `${summary} • Payout impact ${percentLabel}`;
+  }
+
+  updateHint(select.value || '');
+
+  select.addEventListener('change', event => {
+    const nextValue = event.target.value;
+    assignInstanceToNiche(definition.id, instance.id, nextValue || null);
+    updateHint(nextValue || '');
+  });
+
+  container.appendChild(select);
+  container.appendChild(hint);
+
+  return container;
+}
+
+export function emitUIEvent(name) {
+  if (typeof document?.createEvent === 'function') {
+    const event = document.createEvent('Event');
+    event.initEvent(name, true, true);
+    document.dispatchEvent(event);
+    return;
+  }
+  if (typeof Event === 'function') {
+    document.dispatchEvent(new Event(name));
+  }
+}
